+++ environment.yaml
@@ -7,5 +7,3 @@ dependencies:
   - snakemake-minimal=5.4
   - pycountry=20.7.3
   - pip
-  - pip:
-    - -e ./lib
