+++ envs/default.yaml
@@ -20,3 +20,4 @@ dependencies:
     - pip=19.1.1
     - pip:
         - -e ../../lib[geo]
+        - pytest-pythonpath
