+++ pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+norecursedirs = .git .snakemake lib
+python_paths = rules
